PythonApplication1/requirements.txt
azure-common==1.1.4
azure-nspkg==1.0.0
azure-storage==0.32.0
beautifulsoup4==4.5.1
bs4==0.0.1
cffi==1.9.1
idna==2.2
numpy==1.11.2
pandas==0.19.1
pandas-datareader==0.2.1
pyasn1==0.1.9
pycparser==2.17
pypyodbc==1.3.3
python-dateutil==2.6.0
pytz==2016.10
requests==2.5.1
requests-file==1.4.1
six==1.10.0
